Let the resolver look at the kind of input and output for methods and add support for Int16.
R=ager@google.com
BUG=

Review URL: https://codereview.chromium.org/892203005

+++ tests/service_tests/echo/cc/echo_service.h
@@ -13,8 +13,8 @@ class EchoService {
   static void TearDown();
   static int Echo(int n);
   static void EchoAsync(int n, void (*callback)(int));
-  static int Sum(int x, int y);
-  static void SumAsync(int x, int y, void (*callback)(int));
+  static int Sum(short x, int y);
+  static void SumAsync(short x, int y, void (*callback)(int));
 };
 
 #endif  // ECHO_SERVICE_H
