+++ tests/service_tests/echo/objc/echo_service.h
@@ -14,8 +14,8 @@
 + (int)Echo:(int)n;
 + (void)EchoAsync:(int)n withCallback:(void (*)(int))callback;
 + (void)EchoAsync:(int)n withBlock:(void (^)(int))callback;
-+ (int)Sum:(int)x with:(int)y;
-+ (void)SumAsync:(int)x with:(int)y withCallback:(void (*)(int))callback;
-+ (void)SumAsync:(int)x with:(int)y withBlock:(void (^)(int))callback;
++ (int)Sum:(short)x with:(int)y;
++ (void)SumAsync:(short)x with:(int)y withCallback:(void (*)(int))callback;
++ (void)SumAsync:(short)x with:(int)y withBlock:(void (^)(int))callback;
 
 @end
